Add files via upload
Problem Statement: String deletion. In one operation you can pick the characters of same frequency and count it as 1 operation to delete them. Print minimum operations required to delete the complete string.  So for example : if string is "aabbccc", it should print 2 because deletion of a and b takes 1 operation and c takes 1 so it counts to 2. I have implemented using HashMap in Java, still it can be optimized.