Add files via upload
String Deletion Problem: Print the minimum operations required to delete string completely according to the character depending on their frequencies. For eg , if the string is "aabbccc" then print 2 because frequency of "a" and "b" is same so it takes 1 operation and "c" will take another 1 so it counts to 2. The code is using Hash Map but it can be optimized.